+++ source/Maps/Contents/Rivers/RiversTool.ts
@@ -1,5 +1,6 @@
 import { MapObject } from "../../../Model/MapObject";
 import { Point } from "../../../Model/Point";
+import { Vector } from "../../../Model/Vector";
 import { Tool } from "../../../UI/Tools/Tool";
 import { ToolContext } from "../../../UI/Tools/ToolContext";
 import { GridHelper } from "../../../Utilities/GridHelper";
@@ -17,9 +18,10 @@ export class RiversTool implements Tool {
     };
 
     private startPosition?: Point;
-    private activeCell?: CellContext;
+    private previousCell?: CellContext;
+    private previousRiver?: MapObject;
 
-    start(context: ToolContext): void {
+    public start(context: ToolContext) {
         const cell = context.cell;
 
         if (cell === undefined) {
@@ -27,78 +29,53 @@ export class RiversTool implements Tool {
         }
 
         this.startPosition = context.cellPosition;
-        this.activeCell = cell;
+        this.previousCell = cell;
     }
 
-    move(context: ToolContext): void {
-        const activeCell = this.activeCell,
-            cell = context.cell;
-
-        if (this.startPosition === undefined || activeCell === undefined || cell === undefined) {
-            return
-        }
-
-        const position = context.cellPosition;
-
-        if (!GridHelper.cellIsEqual(activeCell.index, cell.index)) {
-            const river = this.getRiver(activeCell)!;
-
-            this.createRivers(activeCell, this.startPosition, cell, position)
-
-            this.startPosition = river.points[0];
-            this.activeCell = cell;
+    public move(context: ToolContext) {
+        if (GridHelper.cellIsEqual(this.previousCell?.index, context.cell?.index)) {
+            this.moveInsideCell(context);
         } else {
-            const river = this.getRiver(cell);
+            this.moveBetweenCells(context);
+        }
+        this.previousCell = context.cell;
+    }
 
-            if (river === undefined) {
-                const from = this.startPosition,
-                    to = position;
+    public stop() {
+        this.startPosition = undefined;
+        this.previousCell = undefined;
+        this.previousRiver = undefined;
+    }
 
-                this.createRiver(cell, from, to);
-            } else {
-                const points = [...river.points];
+    private connectRiver(cell: CellContext, river: MapObject, direction: Vector) {
+        const
+            lastRiverCell = GridHelper.cellNameToIndex(river.cell),
+            neighbor = cell.neighbors.find(c => GridHelper.cellIsEqual(c?.index, lastRiverCell));
 
-                points[1] = position;
-                cell.update(river.id, points);
+        // Last river was in a neighboring cell
+        if (neighbor !== undefined) {
+            const connection = GridHelper.getConnection(river.points[1], direction);
 
-                this.startPosition = river.points[0];
+            // Last river connects to this cell
+            if (neighbor.neighbors[connection.neighborIndex] === cell) {
+                this.updateRiver(neighbor, river, connection.point);
+                return connection;
             }
         }
+
+        return undefined;
     }
 
-    stop(): void {
-        this.startPosition = undefined;
-        this.activeCell = undefined
-    }
+    private createRiver(cell: CellContext, from: Point, to: Point) {
+        const previous = this.previousRiver;
 
-    private createRivers(firstCell: CellContext, start: Point, lastCell: CellContext, end: Point) {
-        const direction = VectorMath.startOperation(end)
-            .add(lastCell.index.column - firstCell.index.column, lastCell.index.row - firstCell.index.row)
-            .direction(start)
-            .invert();
-
-        let cell = firstCell,
-            from = start,
-            previous = this.getRiver(firstCell),
-            [to, nextCell, nextFrom] = GridHelper.getConnection(from, direction),
-            iterations = 0;
-
-        do {
-            cell = cell.neighbors[nextCell]!;
-            from = nextFrom;
-
-            [to, nextCell, nextFrom] = GridHelper.getConnection(from, direction);
-            previous = this.createRiver(cell, from, to, previous);
-            Utilities.checkInfiniteLoop(iterations++);
-        }
-        while (!GridHelper.cellIsEqual(cell.index, lastCell.index));
-    }
-
-    private createRiver(cell: CellContext, from: Point, to: Point, previous?: MapObject) {
         let bend1 = {
             x: MathHelper.random(.2, .8),
             y: MathHelper.random(.2, .8)
         };
+
+        from = VectorMath.round(from, 2);
+        to = VectorMath.round(to, 2);
 
         if (previous !== undefined) {
             bend1 = VectorMath.startOperation(previous.points[3])
@@ -117,10 +94,77 @@ export class RiversTool implements Tool {
         cell.clear();
         cell.addObjects([river]);
 
-        return river;
+        this.previousRiver = river;
     }
 
-    private getRiver(cell: CellContext) {
-        return cell.objects.value.find(o => o.type === 'river');
+    private moveBetweenCells(context: ToolContext) {
+        const cell = context.cell,
+            previousCell = this.previousCell,
+            previousRiver = this.previousRiver,
+            direction = context.direction;
+
+        if (
+            cell === undefined ||
+            previousCell === undefined ||
+            previousRiver === undefined ||
+            !previousCell.hasObject(previousRiver)
+        ) {
+            return;
+        }
+
+        let connection = GridHelper.getConnection(previousRiver.points[1], direction),
+            nextCell = previousCell.neighbors[connection.neighborIndex],
+            iterations = 0;
+
+        this.updateRiver(previousCell, previousRiver, connection.point);
+
+        while (!GridHelper.cellIsEqual(nextCell?.index, cell.index)) {
+            if (nextCell === undefined) {
+                return;
+            }
+
+            const from = connection.neighborPoint;
+
+            connection = GridHelper.getConnection(from, direction);
+
+            this.createRiver(nextCell, from, connection.point);
+
+            nextCell = nextCell.neighbors[connection.neighborIndex];
+
+            Utilities.checkInfiniteLoop(iterations++);
+        }
+
+        if (nextCell !== undefined) {
+            this.createRiver(nextCell, connection.neighborPoint, context.cellPosition);
+        }
+    }
+
+    private moveInsideCell(context: ToolContext) {
+        const cell = context.cell,
+            position = context.cellPosition,
+            lastRiver = this.previousRiver;
+
+        if (cell === undefined || this.startPosition === undefined) {
+            return;
+        }
+
+        if (lastRiver === undefined) {
+            this.createRiver(cell, this.startPosition, position);
+        } else if (cell.hasObject(lastRiver)) {
+            this.updateRiver(cell, lastRiver, position);
+        } else {
+            const connection = this.connectRiver(cell, lastRiver, context.direction),
+                from = connection?.neighborPoint ?? this.startPosition;
+
+            this.createRiver(cell, from, position);
+        }
+
+    }
+
+    private updateRiver(cell: CellContext, river: MapObject, position: Point) {
+        const points = [...river.points];
+
+        points[1] = VectorMath.round(position, 2);
+        cell.update(river.id, points);
     }
 }