+++ source/Model/CellConnection.ts
@@ -1,0 +1,7 @@
+import { Point } from "./Point";
+
+export interface CellConnection {
+    readonly point: Point;
+    readonly neighborIndex: number;
+    readonly neighborPoint: Point;
+}