+++ source/UI/Tools/ToolContext.ts
@@ -1,8 +1,10 @@
 import { CellContext } from "../../Maps/Cells/CellContext";
 import { Point } from "../../Model/Point";
+import { Vector } from "../../Model/Vector";
 
 export interface ToolContext {
-    mapPosition: Point,
+    position: Point,
+    direction: Vector,
     cell?: CellContext,
     cellPosition: Point
 }